requirements.txt
python-telegram-bot[tgcrypto]==21.*
aiosqlite
